+++ src/config_file.rs
@@ -51,6 +51,7 @@ pub struct Color {
 pub struct Icons {
     pub when: Option<IconOption>,
     pub theme: Option<IconTheme>,
+    pub separator: Option<String>,
 }
 
 #[derive(Eq, PartialEq, Debug, Deserialize)]
@@ -191,7 +192,7 @@ color:
 # accepts an strftime like string.
 # When "classic" is set, this is set to "date".
 # Possible values: date, relative, +<date_format>
-date: date
+# date: date
 
 # == Dereference ==
 # Whether to dereference symbolic links.
@@ -212,6 +213,9 @@ icons:
   # Which icon theme to use.
   # Possible values: fancy, unicode
   theme: fancy
+  # The string between the icons and the name.
+  # Possible values: any string (eg: " |")
+  separator: " "
 
 # == Ignore Globs ==
 # A list of globs to ignore when listing.
@@ -301,12 +305,13 @@ mod tests {
                 color: Some(config_file::Color {
                     when: ColorOption::Auto,
                 }),
-                date: Some("date".into()),
+                date: None,
                 dereference: Some(false),
                 display: None,
                 icons: Some(config_file::Icons {
                     when: Some(IconOption::Auto),
                     theme: Some(IconTheme::Fancy),
+                    separator: Some(" ".to_string()),
                 }),
                 ignore_globs: None,
                 indicators: Some(false),