+++ src/display.rs
@@ -373,7 +373,7 @@ mod tests {
             );
             let output = name.render(
                 &Colors::new(color::Theme::NoColor),
-                &Icons::new(icon::Theme::NoIcon),
+                &Icons::new(icon::Theme::NoIcon, " ".to_string()),
                 &DisplayOption::FileName,
             );
 
@@ -405,7 +405,7 @@ mod tests {
             let output = name
                 .render(
                     &Colors::new(color::Theme::NoColor),
-                    &Icons::new(icon::Theme::Fancy),
+                    &Icons::new(icon::Theme::Fancy, " ".to_string()),
                     &DisplayOption::FileName,
                 )
                 .to_string();
@@ -437,7 +437,7 @@ mod tests {
             let output = name
                 .render(
                     &Colors::new(color::Theme::NoLscolors),
-                    &Icons::new(icon::Theme::NoIcon),
+                    &Icons::new(icon::Theme::NoIcon, " ".to_string()),
                     &DisplayOption::FileName,
                 )
                 .to_string();
@@ -473,7 +473,7 @@ mod tests {
             let output = name
                 .render(
                     &Colors::new(color::Theme::NoColor),
-                    &Icons::new(icon::Theme::NoIcon),
+                    &Icons::new(icon::Theme::NoIcon, " ".to_string()),
                     &DisplayOption::FileName,
                 )
                 .to_string();