+++ tests/integration.rs
@@ -106,7 +106,7 @@ fn test_list_inode_populated_directory() {
     #[cfg(windows)]
     let matched = "- one\n\\- two\n$";
     #[cfg(unix)]
-    let matched = "\\d+ one\n\\d+ two\n$";
+    let matched = "\\d+ +one\n\\d+ +two\n$";
 
     cmd()
         .arg("--inode")
@@ -128,7 +128,10 @@ fn test_list_block_inode_populated_directory_without_long() {
     dir.child("one").touch().unwrap();
     dir.child("two").touch().unwrap();
 
-    let matched = "one\ntwo\n$";
+    #[cfg(windows)]
+    let matched = "- one\n\\- two\n$";
+    #[cfg(unix)]
+    let matched = "\\d+ +one\n\\d+ +two\n$";
 
     cmd()
         .arg("--blocks")
@@ -148,7 +151,7 @@ fn test_list_block_inode_populated_directory_with_long() {
     #[cfg(windows)]
     let matched = "- one\n\\- two\n$";
     #[cfg(unix)]
-    let matched = "\\d+ one\n\\d+ two\n$";
+    let matched = "\\d+ +one\n\\d+ +two\n$";
 
     cmd()
         .arg("--long")